Added latest CSS changes and changed component structure for the plant sections
