+++ frontend/src/components/CollectionGrid/CollectionGrid.js
@@ -1,23 +1,24 @@
 import React from 'react';
-import CollectionCard from '../CollectionCard/CollectionCard';
+import CollectionItem from '../CollectionItem/CollectionItem';
 import './CollectionGrid.css';
 
 
 export default function CollectionGrid({ plants, handleNicknameChange, nickname, addToMyPlants }) {
 	return (
-		/* <div className={loading ? 'loaderContainer' : 'grid'}>
+		/* TODO: <div className={loading ? 'loaderContainer' : 'grid'}>
           <Loader></Loader> */
 
 		<div>
-			<ul className='collection__grid'>
+			<ul className='collectionGrid'>
 				{plants.map(plant => (
-					<CollectionCard
+					<CollectionItem
 						handleNicknameChange={handleNicknameChange}
 						nickname={nickname}
 						key={plant._id}
 						id={plant._id}
 						plant={plant}
-						addToMyPlants={addToMyPlants}></CollectionCard>
+						addToMyPlants={addToMyPlants}>
+					</CollectionItem>
 				))}
 			</ul>
 		</div>
