+++ frontend/src/components/CollectionItem/CollectionItem.css
@@ -1,0 +1,46 @@
+.collectionCard {
+	display: flex;
+	flex-direction: column;
+	position: relative;
+	background-color: var(--medium-green);
+	border-radius: 2rem;
+}
+
+.collectionCard:hover {
+	cursor: pointer;
+}
+
+.addIcon {
+	width: 3rem;
+	height: 3rem;
+	font-size: 1.6rem;
+	line-height: 3rem;
+	background-color: var(--lightest);
+	border-radius: 50%;
+	color: var(--medium);
+	z-index: 2;
+	position: relative; 
+	margin:-1.5rem auto 0;
+	text-align:center;
+}
+
+.addIcon:hover {
+	background-color: var(--light-accent);
+	color: var(--lightest);
+}
+
+/* Modal */
+.modalCloseModal {
+	border: 0;
+	background: transparent;
+	color: var(--accent);
+	font-size: 25px;
+	position: absolute;
+	top: 10px;
+	right: 15px;
+}
+
+.modalCloseModal:hover {
+	color: var(--darkest);
+	cursor: pointer;
+}
