+++ frontend/src/components/CollectionModal/CollectionModal.js
@@ -1,4 +1,4 @@
-import './PlantModal.css';
+import './CollectionModal.css';
 
 import SaveToMyPlantsForm from '../SaveToMyPlantsForm/SaveToMyPlantsForm';
 
@@ -10,6 +10,7 @@ import { faSun } from '@fortawesome/free-solid-svg-icons';
 import { faSeedling } from '@fortawesome/free-solid-svg-icons';
 import { faDroplet } from '@fortawesome/free-solid-svg-icons';
 import { faPaw } from '@fortawesome/free-solid-svg-icons';
+import { faTemperatureHalf } from '@fortawesome/free-solid-svg-icons';
 
 const careLevelIcon = <FontAwesomeIcon icon={faLeaf} />;
 const locationInteriorIcon = <FontAwesomeIcon icon={faHouse} />;
@@ -17,8 +18,9 @@ const locationExteriorIcon = <FontAwesomeIcon icon={faSun} />;
 const irrigationIcon = <FontAwesomeIcon icon={faDroplet} />;
 const soilIcon = <FontAwesomeIcon icon={faSeedling} />;
 const petFriendlyIcon = <FontAwesomeIcon icon={faPaw} />;
+const temperatureIcon = <FontAwesomeIcon icon={faTemperatureHalf} />;
 
-export default function PlantModal({
+export default function CollectionModal({
 	id,
 	imageUrl,
 	commonName,
@@ -51,11 +53,11 @@ export default function PlantModal({
 					<img src={`/images/${imageUrl}.jpg`} alt={commonName} />
 					<ul>
 						<li class='card__carelevel'>
-							<h4>Care Level: </h4> {careLevelInfo}
+							<h4>Dificultad: </h4> {careLevelInfo}
 						</li>
 						<li>
 							{petFriendly === true ? (
-								<span className='altIcons' title='Pet Friendly'>
+								<span className='altIcons' title='Apta para mascotas'>
 									{petFriendlyIcon}
 								</span>
 							) : (
@@ -67,64 +69,73 @@ export default function PlantModal({
 				<div class='modalInfo'>
 					<h2>{scientificName}</h2>
 					<h3>{commonName}</h3>
-					{/* <p>{_id}</p> */}
 					<ul>
-						<li class='plantLocation'>
-							<p>
+            			<li class="plantLocation">
+							<div>
 								{interiorExterior !== 1 ? (
-									interiorExterior === 2 ? (
-										<span className='altIcons' title='Exterior'>
-											{locationExteriorIcon}
-										</span>
-									) : (
-										<span className='altIcons' title='Interior &amp; Exterior'>
-											{locationInteriorIcon}
-											{locationExteriorIcon}
-										</span>
-									)
-								) : (
-									<span className='altIcons' title='Interior'>
-										{locationInteriorIcon}
+								interiorExterior === 2 ? (
+									<span className="altIcons" title="Exterior">
+									{locationExteriorIcon}
 									</span>
+								) : (
+									<span className="altIcons double" title="Interior y Exterior">
+									{locationInteriorIcon}
+									{locationExteriorIcon}
+									</span>
+								)
+								) : (
+								<span className="altIcons" title="Interior">
+									{locationInteriorIcon}
+								</span>
 								)}
+							</div>
+							<div>
 								{location}
-							</p>
-						</li>
-						<li class='plantIrrigation'>
-							<span className='altIcons' title='Irrigation'>
-								{irrigationIcon}
+							</div>
+            			</li>
+						<li class="plantIrrigation">
+						<span className="altIcons" title="Riego">
+							{irrigationIcon}
+						</span>
+						<div>
+							<span>
+							{
+								{
+								1: "Una vez al mes",
+								2: "Una vez por semana",
+								3: "2/3 veces a la semana",
+								4: "A diario",
+								}[irrigationSummer]
+							}{" "}
+							in Summer,{" "}
 							</span>
 							<span>
+							{
 								{
-									{
-										1: 'Once a month',
-										2: 'Once a week',
-										3: '2/3 times a week',
-										4: 'Every day',
-									}[irrigationSummer]
-								}{' '}
-								in Summer,{' '}
-							</span>
-							<span>
-								{
-									{
-										1: 'once a month',
-										2: 'once a week',
-										3: '2/3 times a week',
-										4: 'every day',
-									}[irrigationWinter]
-								}{' '}
-								in Winter.
+								1: "una vez al mes",
+								2: "una vez por semana",
+								3: "2/3 veces a la semana",
+								4: "a diario",
+								}[irrigationWinter]
+							}{" "}
+							in Winter.
 							</span>
 							<p>{irrigation}</p>
+						</div>
+						</li>
+						<li className="plantTemp">
+						<span className='altIcons' title='Temperatura'>
+											{temperatureIcon}
+										</span>
+						<div>{temperature}</div>
 						</li>
 						<li>
-							<span className='altIcons' title='Soil'>
-								{soilIcon}
-							</span>
-							{soil}
+						<span className="altIcons" title="Sustrato">
+							{soilIcon}
+						</span>
+						<div>{soil}</div>
 						</li>
-					</ul>
+          			</ul>
 				</div>
 			</div>
 			<SaveToMyPlantsForm
