+++ frontend/src/components/Features/Features.js
@@ -1,6 +1,3 @@
-// Per fer AL FINAL, quan el disseny estigui decidit perquè hi posarem screenshots etc
-
-
 import "./Features.css";
 
 export default function Features() {
