+++ frontend/src/components/MyPlantsGrid/MyPlantsGrid.css
@@ -1,16 +1,55 @@
-.myplants__grid {
-	list-style-type: none;
-	margin: 0;
-	padding: 0;
-	display: grid;
-	grid-template-rows: repeat(3, 1fr);
-	grid-template-columns: repeat(3, 1fr);
-	grid-column-gap: 1rem;
-	grid-row-gap: 1rem;
+.myPlantsHeader {
+	padding: 3rem 0;
+	display: flex;
+	justify-content: space-between;
+	align-items: center;
+	position: relative;
 }
 
-.grid__container {
+.myPlantsHeader .content > * {
+	z-index:1;
+}
+
+.myPlantsHeader::before {
+	content: "";
+	background: url(../../bg_monstera.png);
+	background-repeat: no-repeat;
+	background-attachment: fixed;
+	background-position: 15rem -7rem;
+	background-origin: content-box;
+	background-size: 70%;
+	opacity: 0.2;
+	top: 0;
+	right: 0;
+	bottom:0;
+	left:0;
+	position: absolute;
+  }
+
+
+.myPlantsHeader .content {
 	display: flex;
+	justify-content: space-between;
 	align-items: center;
-	justify-content: center;
+	padding: 2.5rem 0;
+}
+
+.myPlantsHeader h1 {
+	color: var(--lightest);
+	font-family: 'Trocchi', serif;
+	font-size: 5rem;
+	line-height: auto;
+	display: inline-block;
+	flex-grow: 1;
+	text-indent: 2rem;
+}
+
+.myPlantsGrid {
+	list-style-type: none;
+	margin: 0;
+	padding: 4rem 2rem;
+	display:grid;
+	grid-template-columns: repeat(4, 1fr);
+	grid-column-gap: 1.5rem;
+	grid-row-gap: 1.5rem;
 }
