+++ frontend/src/components/MyPlantsGrid/MyPlantsGrid.js
@@ -1,20 +1,21 @@
-import MyPlantsCard from '../MyPlantsCard/MyPlantsCard.js';
+import MyPlantsItem from '../MyPlantsItem/MyPlantsItem.js';
 import './MyPlantsGrid.css';
 
 export default function MyPlantsGrid({ plants, addToMyPlants, deleteFromMyPlants, isLoading }) {
 	console.log(plants);
 	return (
 		<div className='grid__container'>
-			{plants.length === 0 && <span>You have no plants yet</span>}
+			{plants.length === 0 && <span>Todavía no tienes ninguna planta en tu colección</span>}
 			{plants && (
-				<ul className='myplants__grid'>
+				<ul className='myPlantsGrid'>
 					{plants.map(plant => (
-						<MyPlantsCard
+						<MyPlantsItem
 							key={plant.plant._id}
 							nickname={plant.nickname}
 							addToMyPlants={addToMyPlants}
 							deleteFromMyPlants={deleteFromMyPlants}
-							plant={plant.plant}></MyPlantsCard>
+							plant={plant.plant}>
+						</MyPlantsItem >
 					))}
 				</ul>
 			)}
