+++ frontend/src/components/MyPlantsItem/MyPlantsItem.css
@@ -1,0 +1,1 @@
+
