--- frontend/src/components/PlantCard/PlantCard.css
@@ -1,70 +1,0 @@
-.collectionCard .card__petfriendly {
-	position: absolute;
-	top:-1px;
-	right:-1px;
-	border-top-right-radius: 2rem;
-	overflow:hidden;
-	width:6rem;
-	height:6rem;
-	z-index:2;
-	display:block;
-}
-.collectionCard .card__petfriendly span {
-	background-color: var(--accent);
-	display:block;
-	transform:rotate(45deg);
-	text-align:center;
-	color: var(--lightest);
-	width:150%;
-	padding: 0.5rem 0;
-}
-
-
-.collectionCard .card__image {
-	margin:0;
-	padding:0;
-    margin-bottom: -0.8rem;
-}
-
-.collectionCard .card__image img {
-	max-width: 100%;
-	border-top-left-radius: 2rem;
-	border-top-right-radius: 2rem;
-}
-
-.collectionCard .card__info {
-	background-color: var(--medium);
-	border-bottom-left-radius: 2rem;
-	border-bottom-right-radius: 2rem;
-	color: var(--lightest);
-	padding: 1rem;
-	margin: 0;
-	text-align: center;
-    flex-grow: 1;
-}
-
-.collectionCard h2 {
-	margin: 0 0 0.3rem;
-	font-family: 'Poppins', sans-serif;
-	font-weight: bold;
-	font-size: 1.8rem;
-	letter-spacing: 0.1rem;
-}
-
-.collectionCard h3 {
-	margin: 0 0 0.5rem;
-	font-family: 'Poppins', sans-serif;
-	font-weight: normal;
-	font-style:italic;
-	font-size: 1.4rem;
-	max-width: 100%;
-}
-
-.collectionCard .card__carelevel {
-	color: var(--dark);
-	font-size: 2rem;
-}
-
-.collectionCard .card__carelevel i {	
-	margin: 0 0.25rem;
-}
